requirements.txt
# Geração de PDF
reportlab==4.2.2

# Manipulação de dados
pandas==2.2.2

# Manipulação de documentos Word
python-docx==1.1.2

# Manipulação de imagens
Pillow==10.4.0

# Leitura de arquivos Excel (usado pelo pandas)
openpyxl==3.1.5
